==> 2_CLEANER.ipynb <==
import pandas as pd

df = pd.read_csv('raw_scrape.csv')

df.veto_power.value_counts()

df.veto_power = df.veto_power.str.strip()
df['veto_china'] = df.veto_power.str.extract(r"(China)")
df['veto_france'] = df.veto_power.str.extract(r"(France)")
df['veto_russia'] = df.veto_power.str.extract(r"(Russian Federation)")
df['veto_ussr'] = df.veto_power.str.extract(r"(USSR)")
df['veto_uk'] = df.veto_power.str.extract(r"(UK)")
df['veto_usa'] = df.veto_power.str.extract(r"(USA)")

df.veto_power = df.veto_power.str.replace("FranceUK", "France UK")
df.veto_power = df.veto_power.str.replace("FranceUK", "France UK")
df.veto_power = df.veto_power.str.replace("ChinaRussian", "China Russian")
df.veto_power = df.veto_power.str.replace("Russian Federation", "Russian_Federation")

df.veto_power.value_counts()

df.head()

df[df.agenda.isnull()]

#Uh-oh gotta fix that. The first one is on the downing of the MH17 flight over Ukraine.
#The reason my scraper missed it is that the title includes a link to a letter from the Ukranian Ambassador.

df.at[11, 'agenda'] = "Letter dated 28 February 2014 from the Permanent Representative of Ukraine to the United Nations addressed to the President of the Security Council (S/2014/136)"

#The second NaN is about the anexation of Crimea.

df.at[14, 'agenda'] = "Letter dated 28 February 2014 from the Permanent Representative of Ukraine to the United Nations addressed to the President of the Security Council (S/2014/136)"


df.head()

df.to_csv('cleaned_scrape.csv', index=False)

df[df['agenda'].str.contains("Syria")]
