==> 3_EDA.ipynb <==
import pandas as pd

df = pd.read_csv("cleaned_scrape.csv")

print(df[df.agenda.str.contains("Member")].shape)
df_membership = df[df.agenda.str.contains("Member")]

df[df['agenda'].str.contains("Middle East")].shape

df[df['agenda'].str.contains("Palest")].shape

df_membership.veto_power.value_counts()

df[df['agenda'].str.contains("Middle East")].veto_power.value_counts()

df[df['agenda'].str.contains("Palest")].veto_power.value_counts()

#More clearly:
df[df['agenda'].str.contains("Syria")].veto_power.value_counts()

from datetime import datetime

time_str = '16 February 1946'

datetime_obj = datetime.strptime(time_str,"%d %B %Y")

datetime_obj

df.date = df.date.str.replace(r"^(0)", "") #getting rid of the inconsistent formatting for the day.
df.date = df.date.str.replace("14-15", "14") #getting rid of the 2nd day for meetings that ran pas midight. 
df.date = df.date.str.replace("27-28", "27")
df.date = df.date.str.replace("30 January 86", "30 January 1986")
df.date = df.date.str.replace("11-13", "11")
df.date = df.date.str.replace("1980 ", "1980")
df.date = df.date.str.replace("7-9", "7")
df.date = df.date.str.replace("3-4", "3")

def datetime_function(row):
    datetime_obj = datetime.strptime(row.date, "%d %B %Y")
    return datetime_obj

df['datetime'] = df.apply(datetime_function, axis=1)

df.dtypes

#We have datetime! Let's make sure:
print(df.datetime.min(),df.datetime.max())

df['agenda_cleaned'] = df.agenda.str.replace('^The question of ',"")
df['agenda_cleaned'] = df.agenda.str.replace('The situation in ',"")
df['agenda_cleaned'] = df.agenda.str.replace('The situation inMiddle East',"Middle East")
df['agenda_cleaned'] = df.agenda.str.replace('The situation in the Middle East ',"Middle East")
df['agenda_cleaned'] = df.agenda_cleaned.str.replace("^The situation in ", "")
df['agenda_cleaned'] = df.agenda_cleaned.str.replace("^The situation in ", "")
df['agenda_cleaned'] = df.agenda_cleaned.str.replace(" question$", "")
df['agenda_cleaned'] = df.agenda_cleaned.str.replace("^The ", "")
df['agenda_cleaned'] = df.agenda_cleaned.str.replace("^.*Palestinian", "Palestine")
df['agenda_cleaned'] = df.agenda_cleaned.str.replace("^question of ", "")
df['agenda_cleaned'] = df.agenda_cleaned.str.replace('^Middle East ', "Middle East")
df['agenda_cleaned'] = df.agenda_cleaned.str.replace("^the ", "")
df['agenda_cleaned'] = df.agenda_cleaned.str.replace("^Question concerning the situation in", "")
df['agenda_cleaned'] = df.agenda_cleaned.str.replace("^ ", "")
df['agenda_cleaned'] = df.agenda_cleaned.str.replace(" $", "")
df['agenda_cleaned'] = df.agenda_cleaned.str.replace("^Namibia ", "Namibia")
df['agenda_cleaned'] = df.agenda_cleaned.str.replace("Spanish", "Spain")
df['agenda_cleaned'] = df.agenda_cleaned.str.replace("Spanish", "Spain")
df['agenda_cleaned'] = df.agenda_cleaned.str.replace("Complaint by Angola against South Africa", "South Africa")

df.agenda_cleaned.value_counts()

df.groupby('veto_power').agenda_cleaned.value_counts()

df.to_csv("veto_data_cleaned.csv", index=False)
