==> 4_MEMBERSHIP_VETO.ipynb <==
import pandas as pd
import re

df = pd.read_csv('veto_data_cleaned.csv')

df_membership = df[df.agenda_cleaned.str.contains(r"[Mm]embers")]
df_membership[df_membership.veto_china == 'China']

df_membership.agenda_cleaned

#Hmm, the best way to clean this would be to use some str.replace, it's too irregular for split(' ') and then indexing.

df_membership['impacted_country'] = df_membership.agenda_cleaned.str.replace('Application for Membership ','')

df_membership.impacted_country = df_membership.impacted_country.str.replace('Admission of new Members ','')
df_membership.impacted_country = df_membership.impacted_country.str.replace('Applications for [Mm]embership ','')
df_membership.impacted_country = df_membership.impacted_country.str.replace(".*Viet-Nam", 'Viet Nam')
df_membership.impacted_country = df_membership.impacted_country.str.replace("Viet Nam", 'Vietnam')
#Let's change Ceylon to Sri Lanka to get the right shapefile:
df_membership.impacted_country = df_membership.impacted_country.str.replace('Ceylon', 'Sri Lanka')
#And Transjordan to Jordan.
df_membership.impacted_country = df_membership.impacted_country.str.replace('Transjordan', 'Jordan')
#And South Korea
df_membership.impacted_country = df_membership.impacted_country.str.replace('Republic of Korea', 'South Korea')

df_membership.impacted_country.value_counts()

df_membership.veto_power.value_counts()

#Let's change these to what a map can handle:
df_membership.veto_power = df_membership.veto_power.str.replace('USSR','Russia')
df_membership.veto_power = df_membership.veto_power.str.replace('USA','United States of America')

df_membership.to_csv('membership_veto.csv', index=False)
