==> 5_LEAFLETTING.ipynb <==
import requests
import json
import numpy as np
import pandas as pd
from pandas.io.json import json_normalize

with open('vetoes_countries.json') as json_data:
    geometry_data = json.load(json_data)
for feat in geometry_data['features']:
    del feat["properties"]['brk_group']
for feat in geometry_data['features']:
    del feat["properties"]['formal_fr']
for feat in geometry_data['features']:
    del feat["properties"]['note_adm0']
for feat in geometry_data['features']:
    del feat["properties"]['note_brk']
for feat in geometry_data['features']:
    del feat["properties"]['name_alt']

df = pd.DataFrame.from_dict(json_normalize(geometry_data['features']), orient='columns')

#Ugh, uniform country names guys. I'll have to fix it so it matches my csv.
df['properties.name'] = df['properties.name'].str.replace('Lao PDR', 'Laos')
df['properties.name'] = df['properties.name'].str.replace('Korea', 'South Korea')

#Let's pull in our csv data
df2 = pd.read_csv('membership_veto.csv')

#Let's join 'em up
output = pd.merge(df2, df[['geometry.coordinates','geometry.type','properties.name']], left_on='impacted_country', right_on='properties.name', how='left')

output['properties.headline'] = output['properties.name']
output['properties.color'] = '#CC3333'
output['properties.article'] = output['properties.name'] + ' had their application to the UN vetoed by ' + output['veto_power']
output['properties.group_name'] = output['veto_power']
output['properties.group_id'] = 0

output['properties.group_name']

output.loc[0:5, 'properties.group_id'] = 2
output.loc[6, 'properties.group_id'] = 3
output.loc[7:, 'properties.group_id'] = 1

output

def fixing_article(row):
    #print('----------------')
    if output.impacted_country.value_counts()[row.impacted_country] == 1:
        #print('1')
        draft_resolution = "Draft resolution: <a href=" + row.draft_url + ">" + row.draft + "</a><br>"
        meeting_notes = "Meeting notes: <a href=" + row.meeting_record_url+ ">" + row.meeting_record + "</a><br>"
        agenda_item = "Agenda item: " + row.agenda + "<br>"
        row['properties.article'] = "<b>" + row.impacted_country + " had their application for membership vetoed on " + row.date + " by " + row.veto_power + ".</b><br>" + draft_resolution + meeting_notes + agenda_item
    if output.impacted_country.value_counts()[row.impacted_country] > 1:
        #print("several"+row.impacted_country+row.draft)
        veto_number = output.impacted_country.value_counts()[row.impacted_country]
        #print(veto_number)
        article_string = ''
        for index, row_inner in output[output.impacted_country == row.impacted_country].iterrows():
            draft_resolution = "Draft resolution: <a href=" + row_inner.draft_url + ">" + row_inner.draft + "</a><br>"
            meeting_notes = "Meeting notes: <a href=" + row_inner.meeting_record_url+ ">" + row_inner.meeting_record + "</a><br>"
            agenda_item = "Agenda item: " + row_inner.agenda + "<br>"
            article_string += "<b>On " + row_inner.date + " by " + row_inner.veto_power + ".</b><br>" + draft_resolution + meeting_notes + agenda_item
        row['properties.article'] = row.impacted_country + " had their application for membership vetoed <b>" + str(veto_number) + "</b> times:<br>" + article_string
    return row

df_final = output.apply(fixing_article, axis=1)

ok_json = json.loads(df_final.to_json(orient='records'))

def process_to_geojson(file):
    geo_data = {"type": "FeatureCollection", "features":[]}
    for row in file:
        this_dict = {"type": "Feature", "properties":{}, "geometry": {}}
        for key, value in row.items():
            key_names = key.split('.')
            if key_names[0] == 'geometry':
                this_dict['geometry'][key_names[1]] = value
            if str(key_names[0]) == 'properties':
                this_dict['properties'][key_names[1]] = value
        geo_data['features'].append(this_dict)
    return geo_data

geo_format = process_to_geojson(ok_json)

with open('geo-data.js', 'w') as outfile:
    json.dump(geo_format, outfile)

df_final.to_csv('membership_veto_final.csv', index=False)
